refactor(db): isolate connection write coordination
